refactor(core): extract workflow support owner
