+++ crates/tokmd-core/src/workflows/module.rs
@@ -5,9 +5,11 @@ use tokmd_settings::ScanOptions;
 use tokmd_types::ModuleReceipt;
 
 use crate::settings::{ModuleSettings, ScanSettings};
-use crate::{
-    InMemoryFile, build_module_receipt, collect_pure_in_memory_rows,
-    deterministic_in_memory_scan_options, scan_paths_or_current_dir, settings_to_scan_options,
+use crate::{InMemoryFile, build_module_receipt};
+
+use super::{
+    collect_pure_in_memory_rows, deterministic_in_memory_scan_options, scan_paths_or_current_dir,
+    settings_to_scan_options,
 };
 
 /// Runs the module summary workflow with pure settings types.
